requirements.txt
azure-ai-inference>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
faker>=19.0.0
streamlit>=1.28.0
plotly>=5.15.0
altair>=5.0.0
streamlit-option-menu>=0.3.6

# Dashboard specific dependencies
plotly>=5.15.0  # For advanced charts and visualizations
